Reorder definitions, add integration test
There are other various Go module and workspace fixes, and coverage
report to work correctly. When the combined coverage report is wanted,
we need to make use of go.work and is not a simple task (as it would
need to place a go.work with all the modules in place, regardless of
what other coding tasks are happening). Because checking the coverage
for overall modules is rather rare as long as we can stick to 100%
test coverage per module, I'm just keeping it around for now, which we
may be able to delete or potentially improve by providing better
command to handle such case.